refactor events to separate file
